Update network-mode env var to be consistent with other CLI flags.
The network-mode CLI flag had an env var set to
PLUGIN_TASK_NETWORK_MODE which broken convention with all other
CLI flags and made it confusing to switch between using drone.yaml
config files and running the plugin differently.

This change updates the env var to be consistent so the user
experience is smoother.